+++ src/main/resources/db/migration/V6__create_table_account_user.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS sch_user.account_user (
     password VARCHAR(255) NOT NULL,
     role     VARCHAR(30) NOT NULL,
     enabled BOOLEAN NOT NULL DEFAULT TRUE,
-    person_id INTEGER NOT NULL,
+    person_id BIGSERIAL NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
     FOREIGN KEY (person_id) REFERENCES sch_person.person(id)
